requirements.txt
numpy
tqdm
glog
easydict
tensorflow_gpu
matplotlib
opencv
scikit_learn
loguru
